Move all our custom object implementations into a separate module, and use a macro to implement them.
