+++ core/src/avm2.rs
@@ -28,15 +28,12 @@ mod function;
 mod globals;
 mod method;
 mod names;
-mod namespace_object;
 mod object;
-mod primitive_object;
 mod property;
 mod property_map;
 mod return_value;
 mod scope;
 mod script;
-mod script_object;
 mod slot;
 mod string;
 mod r#trait;
