+++ core/src/avm2/activation.rs
@@ -1,14 +1,12 @@
 //! Activation frames
 
 use crate::avm2::class::Class;
-use crate::avm2::function::FunctionObject;
 use crate::avm2::method::BytecodeMethod;
 use crate::avm2::names::{Multiname, Namespace, QName};
-use crate::avm2::namespace_object::NamespaceObject;
+use crate::avm2::object::{FunctionObject, NamespaceObject, ScriptObject};
 use crate::avm2::object::{Object, TObject};
 use crate::avm2::scope::Scope;
 use crate::avm2::script::Script;
-use crate::avm2::script_object::ScriptObject;
 use crate::avm2::string::AvmString;
 use crate::avm2::value::Value;
 use crate::avm2::{value, Avm2, Error};
