+++ core/src/avm2/globals/flash/display/displayobjectcontainer.rs
@@ -1,8 +1,7 @@
 //! `flash.display.DisplayObjectContainer` builtin/prototype
 
 use crate::avm2::activation::Activation;
-use crate::avm2::object::Object;
-use crate::avm2::script_object::ScriptObject;
+use crate::avm2::object::{Object, ScriptObject};
 use crate::avm2::value::Value;
 use crate::avm2::Error;
 use gc_arena::MutationContext;
