--- core/src/avm2/namespace_object.rs
@@ -1,354 +1,0 @@
-//! Boxed namespaces
-
-use crate::avm1::AvmString;
-use crate::avm2::activation::Activation;
-use crate::avm2::class::Class;
-use crate::avm2::function::Executable;
-use crate::avm2::names::{Namespace, QName};
-use crate::avm2::object::{Object, ObjectPtr, TObject};
-use crate::avm2::r#trait::Trait;
-use crate::avm2::scope::Scope;
-use crate::avm2::script_object::{ScriptObjectClass, ScriptObjectData};
-use crate::avm2::value::Value;
-use crate::avm2::Error;
-use gc_arena::{Collect, GcCell, MutationContext};
-use std::cell::Ref;
-
-/// An Object which represents a boxed namespace name.
-#[derive(Collect, Debug, Clone, Copy)]
-#[collect(no_drop)]
-pub struct NamespaceObject<'gc>(GcCell<'gc, NamespaceObjectData<'gc>>);
-
-#[derive(Collect, Debug, Clone)]
-#[collect(no_drop)]
-pub struct NamespaceObjectData<'gc> {
-    /// All normal script data.
-    base: ScriptObjectData<'gc>,
-
-    /// The namespace name this object is associated with.
-    namespace: Namespace<'gc>,
-}
-
-impl<'gc> NamespaceObject<'gc> {
-    /// Box a namespace into an object.
-    pub fn from_namespace(
-        namespace: Namespace<'gc>,
-        base_proto: Object<'gc>,
-        mc: MutationContext<'gc, '_>,
-    ) -> Result<Object<'gc>, Error> {
-        let base = ScriptObjectData::base_new(Some(base_proto), ScriptObjectClass::NoClass);
-
-        Ok(NamespaceObject(GcCell::allocate(
-            mc,
-            NamespaceObjectData { base, namespace },
-        ))
-        .into())
-    }
-}
-
-impl<'gc> TObject<'gc> for NamespaceObject<'gc> {
-    fn get_property_local(
-        self,
-        reciever: Object<'gc>,
-        name: &QName<'gc>,
-        activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Result<Value<'gc>, Error> {
-        let read = self.0.read();
-        let rv = read.base.get_property_local(reciever, name, activation)?;
-
-        drop(read);
-
-        rv.resolve(activation)
-    }
-
-    fn set_property_local(
-        self,
-        reciever: Object<'gc>,
-        name: &QName<'gc>,
-        value: Value<'gc>,
-        activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Result<(), Error> {
-        let mut write = self.0.write(activation.context.gc_context);
-        let rv = write
-            .base
-            .set_property_local(reciever, name, value, activation)?;
-
-        drop(write);
-
-        rv.resolve(activation)?;
-
-        Ok(())
-    }
-
-    fn init_property_local(
-        self,
-        reciever: Object<'gc>,
-        name: &QName<'gc>,
-        value: Value<'gc>,
-        activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Result<(), Error> {
-        let mut write = self.0.write(activation.context.gc_context);
-        let rv = write
-            .base
-            .init_property_local(reciever, name, value, activation)?;
-
-        drop(write);
-
-        rv.resolve(activation)?;
-
-        Ok(())
-    }
-
-    fn is_property_overwritable(
-        self,
-        gc_context: MutationContext<'gc, '_>,
-        name: &QName<'gc>,
-    ) -> bool {
-        self.0.write(gc_context).base.is_property_overwritable(name)
-    }
-
-    fn delete_property(
-        &self,
-        gc_context: MutationContext<'gc, '_>,
-        multiname: &QName<'gc>,
-    ) -> bool {
-        self.0.write(gc_context).base.delete_property(multiname)
-    }
-
-    fn get_slot(self, id: u32) -> Result<Value<'gc>, Error> {
-        self.0.read().base.get_slot(id)
-    }
-
-    fn set_slot(
-        self,
-        id: u32,
-        value: Value<'gc>,
-        mc: MutationContext<'gc, '_>,
-    ) -> Result<(), Error> {
-        self.0.write(mc).base.set_slot(id, value, mc)
-    }
-
-    fn init_slot(
-        self,
-        id: u32,
-        value: Value<'gc>,
-        mc: MutationContext<'gc, '_>,
-    ) -> Result<(), Error> {
-        self.0.write(mc).base.init_slot(id, value, mc)
-    }
-
-    fn get_method(self, id: u32) -> Option<Object<'gc>> {
-        self.0.read().base.get_method(id)
-    }
-
-    fn get_trait(self, name: &QName<'gc>) -> Result<Vec<Trait<'gc>>, Error> {
-        self.0.read().base.get_trait(name)
-    }
-
-    fn get_provided_trait(
-        &self,
-        name: &QName<'gc>,
-        known_traits: &mut Vec<Trait<'gc>>,
-    ) -> Result<(), Error> {
-        self.0.read().base.get_provided_trait(name, known_traits)
-    }
-
-    fn get_scope(self) -> Option<GcCell<'gc, Scope<'gc>>> {
-        self.0.read().base.get_scope()
-    }
-
-    fn resolve_any(self, local_name: AvmString<'gc>) -> Result<Option<Namespace<'gc>>, Error> {
-        self.0.read().base.resolve_any(local_name)
-    }
-
-    fn resolve_any_trait(
-        self,
-        local_name: AvmString<'gc>,
-    ) -> Result<Option<Namespace<'gc>>, Error> {
-        self.0.read().base.resolve_any_trait(local_name)
-    }
-
-    fn has_own_property(self, name: &QName<'gc>) -> Result<bool, Error> {
-        self.0.read().base.has_own_property(name)
-    }
-
-    fn has_trait(self, name: &QName<'gc>) -> Result<bool, Error> {
-        self.0.read().base.has_trait(name)
-    }
-
-    fn provides_trait(self, name: &QName<'gc>) -> Result<bool, Error> {
-        self.0.read().base.provides_trait(name)
-    }
-
-    fn has_instantiated_property(self, name: &QName<'gc>) -> bool {
-        self.0.read().base.has_instantiated_property(name)
-    }
-
-    fn has_own_virtual_getter(self, name: &QName<'gc>) -> bool {
-        self.0.read().base.has_own_virtual_getter(name)
-    }
-
-    fn has_own_virtual_setter(self, name: &QName<'gc>) -> bool {
-        self.0.read().base.has_own_virtual_setter(name)
-    }
-
-    fn proto(&self) -> Option<Object<'gc>> {
-        self.0.read().base.proto()
-    }
-
-    fn get_enumerant_name(&self, index: u32) -> Option<QName<'gc>> {
-        self.0.read().base.get_enumerant_name(index)
-    }
-
-    fn property_is_enumerable(&self, name: &QName<'gc>) -> bool {
-        self.0.read().base.property_is_enumerable(name)
-    }
-
-    fn set_local_property_is_enumerable(
-        &self,
-        mc: MutationContext<'gc, '_>,
-        name: &QName<'gc>,
-        is_enumerable: bool,
-    ) -> Result<(), Error> {
-        self.0
-            .write(mc)
-            .base
-            .set_local_property_is_enumerable(name, is_enumerable)
-    }
-
-    fn as_ptr(&self) -> *const ObjectPtr {
-        self.0.as_ptr() as *const ObjectPtr
-    }
-
-    fn as_executable(&self) -> Option<Executable<'gc>> {
-        None
-    }
-
-    fn construct(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        _args: &[Value<'gc>],
-    ) -> Result<Object<'gc>, Error> {
-        let this: Object<'gc> = Object::NamespaceObject(*self);
-        let base = ScriptObjectData::base_new(Some(this), ScriptObjectClass::NoClass);
-
-        Ok(NamespaceObject(GcCell::allocate(
-            activation.context.gc_context,
-            NamespaceObjectData {
-                base,
-                namespace: Namespace::public_namespace(),
-            },
-        ))
-        .into())
-    }
-
-    fn derive(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        class: GcCell<'gc, Class<'gc>>,
-        scope: Option<GcCell<'gc, Scope<'gc>>>,
-    ) -> Result<Object<'gc>, Error> {
-        let this: Object<'gc> = Object::NamespaceObject(*self);
-        let base = ScriptObjectData::base_new(
-            Some(this),
-            ScriptObjectClass::InstancePrototype(class, scope),
-        );
-
-        Ok(NamespaceObject(GcCell::allocate(
-            activation.context.gc_context,
-            NamespaceObjectData {
-                base,
-                namespace: Namespace::public_namespace(),
-            },
-        ))
-        .into())
-    }
-
-    fn to_string(&self, _mc: MutationContext<'gc, '_>) -> Result<Value<'gc>, Error> {
-        Ok(self.0.read().namespace.as_uri().into())
-    }
-
-    fn value_of(&self, _mc: MutationContext<'gc, '_>) -> Result<Value<'gc>, Error> {
-        Ok(self.0.read().namespace.as_uri().into())
-    }
-
-    fn install_method(
-        &mut self,
-        mc: MutationContext<'gc, '_>,
-        name: QName<'gc>,
-        disp_id: u32,
-        function: Object<'gc>,
-    ) {
-        self.0
-            .write(mc)
-            .base
-            .install_method(name, disp_id, function)
-    }
-
-    fn install_getter(
-        &mut self,
-        mc: MutationContext<'gc, '_>,
-        name: QName<'gc>,
-        disp_id: u32,
-        function: Object<'gc>,
-    ) -> Result<(), Error> {
-        self.0
-            .write(mc)
-            .base
-            .install_getter(name, disp_id, function)
-    }
-
-    fn install_setter(
-        &mut self,
-        mc: MutationContext<'gc, '_>,
-        name: QName<'gc>,
-        disp_id: u32,
-        function: Object<'gc>,
-    ) -> Result<(), Error> {
-        self.0
-            .write(mc)
-            .base
-            .install_setter(name, disp_id, function)
-    }
-
-    fn install_dynamic_property(
-        &mut self,
-        mc: MutationContext<'gc, '_>,
-        name: QName<'gc>,
-        value: Value<'gc>,
-    ) -> Result<(), Error> {
-        self.0.write(mc).base.install_dynamic_property(name, value)
-    }
-
-    fn install_slot(
-        &mut self,
-        mc: MutationContext<'gc, '_>,
-        name: QName<'gc>,
-        id: u32,
-        value: Value<'gc>,
-    ) {
-        self.0.write(mc).base.install_slot(name, id, value)
-    }
-
-    fn install_const(
-        &mut self,
-        mc: MutationContext<'gc, '_>,
-        name: QName<'gc>,
-        id: u32,
-        value: Value<'gc>,
-    ) {
-        self.0.write(mc).base.install_const(name, id, value)
-    }
-
-    fn interfaces(&self) -> Vec<Object<'gc>> {
-        self.0.read().base.interfaces()
-    }
-
-    fn set_interfaces(&self, context: MutationContext<'gc, '_>, iface_list: Vec<Object<'gc>>) {
-        self.0.write(context).base.set_interfaces(iface_list)
-    }
-
-    fn as_namespace(&self) -> Option<Ref<Namespace<'gc>>> {
-        Some(Ref::map(self.0.read(), |s| &s.namespace))
-    }
-}
