+++ core/src/avm2/object/script_object.rs
